Fixed 1080p support

+++ MyGame/Buttons.h
@@ -13,7 +13,7 @@ public:
         SDL_SetRenderDrawColor(r, color[0], color[1], color[2], color[3]);
         SDL_RenderFillRect(r, &buttonRect);
     }
-    bool clicked(int scale) {
+    bool clicked(float scale) {
         float x, y;
         SDL_GetMouseState(&x, &y);
         if (x >= buttonRect.x * scale && x <= (buttonRect.w + buttonRect.x) * scale && y >= buttonRect.y * scale && y <= (buttonRect.h + buttonRect.y) * scale) {
